+++ src/mdl_g2c/g2clib-1.0.5/g2_info.c
@@ -178,7 +178,8 @@ g2int g2_info(unsigned char *cgrib,g2int *listsec0,g2int *listsec1,
               (*numfields)++;
         }
         else {
-           printf("g2_info: Invalid section number found in GRIB message: %ld\n"                   ,isecnum);
+           printf("g2_info: Invalid section number found in GRIB message: %d\n",
+                  (int)isecnum);
            ierr=6;
            return(ierr);
         }
